Check for directory existence and create if absent
For multi-lingual books (e.g. book/en, book/fr, etc.) if you try and
build styles written in LESS it will fail because the init hook
triggers before the language directories are written. This change
checks to see if the path exists first and, if it doesn't, creates it
synchronously.